LPS-100020 extract code to Logic class to be called externally by the headless APIs
